+++ src/routes/home/Home.module.css
@@ -149,6 +149,8 @@ h3 {
 .bigItemList li {
   display: inline-block;
   width: 19.4%;
+  max-width: 197px;
+  min-width: 150px;
   height: auto;
 }
 
@@ -158,6 +160,12 @@ h3 {
 }
 
 .smallListContainer {
+  position: relative;
+  height: auto;
+  width: 100%;
+}
+
+.categoryListContainer {
   position: relative;
   height: auto;
   width: 100%;
@@ -191,6 +199,40 @@ h3 {
 .smallItemList li {
   flex-shrink: 0;
   width: 15.3%;
+  max-width: 157px;
+  min-width: 127px;
+  height: auto;
+}
+
+.categoryItemList {
+  display: flex;
+  justify-content: space-between;
+  list-style: none;
+  margin-top: 26px;
+
+  overflow-y: hidden;
+  overflow-x: scroll;
+  scroll-behavior: smooth;
+
+  grid-gap: 1.65%;
+  /* grid-gap: 17px; */
+  transition-delay: 1s;
+
+  margin-left: -15px;
+  margin-right: -15px;
+  padding: 4px 15px;
+
+}
+
+.categoryItemList::-webkit-scrollbar {
+  display: none;
+}
+
+.categoryItemList li {
+  flex-shrink: 0;
+  width: 15.3%;
+  max-width: 157px;
+  min-width: 127px;
   height: auto;
 }
 
@@ -244,7 +286,7 @@ h3 {
   }
 }
 */
-@media screen and (max-width: 960px) {
+@media screen and (max-width: 900px) {
   h2 {
     font-size: 28px;
     line-height: 42px;
@@ -262,9 +304,7 @@ h3 {
   }
 
   .homeCategoryContainer {
-    width: 100%;
-    height: 116px;
-    position: relative;
+    height: 100px;
   }
   
   .homeCategoryNav2 {
@@ -278,6 +318,10 @@ h3 {
     cursor: pointer;
     width: 67px;
     height: 67px;
+  }
+
+  .todaysSpecial {
+    padding: 20px 0 22px 0;
   }
 
   .homeBulletin {
@@ -298,39 +342,15 @@ h3 {
   }
 
   .bigItemList {
-    padding: 4px 0;
-    list-style: none;
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-
-    margin-left: auto;
-    margin-right: auto;
-
-    width: 300px;
-    min-width: 300px;
-
-    grid-column-gap: 20px;
+    
   }
 
   .bigItemList li {
-    width: 140px;
+    width: 150px;
     height: auto;
     padding: 16px 0;
   }
 
-  .smallItemList {
-    overflow-y: hidden;
-    overflow-x: scroll;
-    /* padding-left: 15px;
-    padding-right: 15px; */
-    margin-left: -15px;
-    margin-right: -15px;
-    padding: 4px 15px;
-  }
-
-  .smallItemList::-webkit-scrollbar {
-    display: none;
-  }
 
   .smallItemSection h3 {
     font-size: 22px;
@@ -346,4 +366,25 @@ h3 {
     width: 70px;
     height: 70px;
   } 
+
+  .categoryItemList {
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    justify-content: space-between;
+  }
+  
+  .categoryItemList li {
+    width: 108px;
+  }
+}
+
+@media screen and (max-width: 700px) {
+  .bigItemList {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    margin-left: auto;
+    margin-right: auto;
+    grid-gap: 10px;
+    width: 310px;
+  }
 }
