+++ src/shared/bottom-nav/muiStyles.js
@@ -1,0 +1,25 @@
+export const bottomNavStyles = {
+
+    bottomNav: {
+      width: "100%",
+      bottom: 0,
+      height: "auto",
+      padding: "0.1% 0",
+      maxHeight: 70,
+      borderTop: "2px solid #B8B8B8",
+      position: "sticky"
+    },
+  
+    bottomNavButton: {
+        width: "100%",
+  
+        "&:hover": {
+            color: "#7141fa"
+        },
+    },
+  
+    buttonNavIcon: {
+        height: "auto",
+        width: "22%"
+    }
+  };
