+++ src/shared/view-cart/ViewCart.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { viewCartStyles } from '../../muiStyles.js';
+import { viewCartStyles, headers } from './muiStyles.js';
 
 /*
  * Material-UI Imports
@@ -13,17 +13,17 @@ function ViewCart({numItems, totalAmount}) {
     <Button sx={viewCartStyles.viewCartContainer}>
         <Grid container spacing={0} sx={viewCartStyles.viewCartGrid}>
             <Grid item xs={3.5} sx={viewCartStyles.viewCartText}>
-                <Typography variant="h5" color="secondary.main" fontWeight={600} marginBottom="6px">
+                <Typography sx={headers.header5} fontWeight={600} marginBottom="6px">
                     {numItems} Items
                 </Typography>
             </Grid>
             <Grid item xs={5} sx={viewCartStyles.viewCartText}>
-                <Typography variant="h3" color="secondary.main" fontWeight={600}>
+                <Typography sx={headers.header3} fontWeight={600}>
                     View Cart
                 </Typography>
             </Grid>
             <Grid item xs={3.5} sx={viewCartStyles.viewCartText}>
-                <Typography variant="h5" color="secondary.main" fontWeight={600} marginBottom="6px">
+                <Typography sx={headers.header5} fontWeight={600} marginBottom="6px">
                     ${totalAmount}
                 </Typography>
             </Grid>
