+++ src/shared/view-cart/muiStyles.js
@@ -1,0 +1,47 @@
+export const headers = {
+
+    header3: {
+        fontWeight: 700,
+        fontSize: "24px",
+        color: "#7141FA"
+      },
+
+    header5: {
+        fontWeight: 700,
+        fontSize: "18px",
+        lineHeight: "21px",
+        color: "#7141FA"
+    }
+
+}
+
+export const viewCartStyles = {
+
+    viewCartContainer: {
+        position: "sticky",
+        left: "50%",
+        transform: "translateX(-50%)",
+        bottom: "115px",
+        width: 384,
+        height: 60,
+        backgroundColor: "#ffffff",
+        border: "1px solid #C4C4C4",
+        borderRadius: "20px",
+        textTransform: 'none',
+        "&:hover": {
+            backgroundColor: "#ffffff"
+        },
+        "@media screen and (max-width: 900px)": {
+            display: "none"
+        }
+    },
+
+    viewCartGrid: {
+        direction: "column",
+        alignItems: "flex-end"
+    },
+
+    viewCartText: {
+        textAlign: "center",
+    }
+}
